+++ server/api.js
@@ -2,7 +2,8 @@ import { Router } from "express";
 import logger from "./utils/logger";
 import db from "./db"
 import getData from "./Controllers/getData";
-import { getAllChannels } from "./slackMethods";
+import postChannel from "./Controllers/postChannel";
+import getChannels from "./Controllers/getChannels";
 
 const router = Router();
 router.get("/", (_, res) => {
@@ -11,25 +12,7 @@ router.get("/", (_, res) => {
 });
 
 router.get("/data", getData);
+router.get("/channels", getChannels)
+router.post("/channel", postChannel )
 
-router.get("/channels", async(req, res) => {
-	let data = await db.query(`SELECT * FROM public.channel`)
-	data = data.rows.map((el) => el.channel_name)
-	res.json(data)
-})
-
-router.post("/channel", async(req, res) => {
-	const { channels } = await getAllChannels();
-	const filteredChannels = channels.filter(
-		(channel) => req.body.data === channel.name
-	);
-	const id = filteredChannels[0].id;
-	const channelName = filteredChannels[0].name;
-
-	const query = `INSERT INTO public.channel(channel_id, channel_name)
-										VALUES ($1, $2);`;
-
-  await db.query(query, [id, channelName]);
-	res.send({status:"success", data: channelName});
-})
 export default router;
